astrocytes-services-py/astrocytes-rest/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astrocytes-rest"
version = "0.1.0"
description = "Reference HTTP server for the Astrocytes memory framework (REST over astrocytes-py)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "astrocytes==0.1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]
pgvector = ["astrocytes-pgvector"]

# `astrocytes` is not published to PyPI yet; resolve it from the sibling package when using uv.
[tool.uv.sources]
astrocytes = { path = "../../astrocytes-py", editable = true }
astrocytes-pgvector = { path = "../astrocytes-pgvector", editable = true }

[project.scripts]
astrocytes-rest = "astrocytes_rest.main:main"

[tool.hatch.build.targets.wheel]
packages = ["astrocytes_rest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
